feat: Move getLinkedData method to utils folder
