+++ src/plugins/policyPack/index.ts
@@ -15,6 +15,7 @@ import { Result, Rule, Severity } from '../../rules-engine/types'
 import Plugin, { ConfiguredPlugin, PluginManager } from '../types'
 import DgraphDataProcessor from '../../rules-engine/data-processors/dgraph-data-processor'
 import DataProcessor from '../../rules-engine/data-processors/data-processor'
+import getLinkedData from '../../utils/data'
 
 export default class PolicyPackPlugin extends Plugin {
   constructor({
@@ -272,60 +273,6 @@ export default class PolicyPackPlugin extends Plugin {
     }
   }
 
-  private getLinkedData({ providerData }: { providerData: ProviderData }): any {
-    const linkedData = {}
-    const allEntities = providerData?.entities || []
-    const allConnections = providerData?.connections || {}
-    const entitiesById: { [key: string]: any } = {}
-
-    for (const entity of allEntities) {
-      // AddawsEc2Input! => queryawsEc2
-      const mutationName = /(?<=\[)(.*?)(?=\])/
-        .exec(entity.mutation as any)[0]
-        .replace('Add', 'query')
-        .replace('Input!', '')
-
-      linkedData[mutationName] = entity.data
-
-      for (const entityData of entity.data) {
-        entitiesById[entityData.id] = entityData
-        // eslint-disable-next-line no-underscore-dangle
-        entityData.__typename = mutationName.replace('query', '') // or entity.name?
-      }
-    }
-
-    // connect data on a second pass
-    for (const entityId of Object.keys(allConnections)) {
-      const entityConnections = allConnections[entityId]
-      const entity = entitiesById[entityId]
-      if (!entity) {
-        // eslint-disable-next-line no-continue
-        continue
-      }
-      for (const conn of entityConnections) {
-        const targetEntity = entitiesById[conn.id]
-        if (!targetEntity) {
-          // eslint-disable-next-line no-continue
-          continue
-        }
-        if (conn.relation === 'child') {
-          if (!entity[conn.field]) {
-            entity[conn.field] = []
-          }
-          entity[conn.field].push(targetEntity)
-          // inverse relation
-          // const inverseConnField = this.schemasMap[entity.__typename] || 'account' // @TODO: account doesn't have a name
-          // if (!targetEntity[inverseConnField]) {
-          //   targetEntity[inverseConnField] = []
-          // }
-          // targetEntity[inverseConnField].push(entity)
-        } // else parent relation.. is not used atm
-      }
-    }
-
-    return linkedData
-  }
-
   async execute({
     storageRunning,
     storageEngine,
@@ -368,7 +315,8 @@ export default class PolicyPackPlugin extends Plugin {
           mergeSchemas(currentSchema, findingsSchema),
         ])
 
-        const linkedData = this.getLinkedData({ providerData })
+        // Format metadata and link connections
+        const linkedData = getLinkedData(providerData)
 
         const findings = await this.executeRule({
           data: linkedData,
